First assignment submission
CEN 524 ASSIGNMENT 
Students should write a python function for each algorithm. Code will be checked 
on April 2nd whilst algorithm will be uploaded to moodle next week. 
~ Osemwegie Omoruyi 
TASKS 
1. Write an algorithm for summing of the Taylor series.        
(Hint: f(x) = f(a)(x − a) + [f '(a)/2! .(x−a). 2] + [f ''(a)/3! .(x − a) .3] + ….. + [f n(a)/n! .(x − a) 
n];    let f(a), f '(a)….. f n(a) be an array of input terms and a is a real number) 
2. With an algorithm, write out the procedure for calculating the factorial of a 
given integer n; (i.e. n!). 
3. With the aid of an algorithm, write out the procedure for calculating the 
Pearson’s Correlation Coefficient of two variables (x and y)  
�
� = ∑(𝑥− 𝑥)(𝑦− 𝑦̅)
 ∑(𝑥 − 𝑥)2∑(𝑦− 𝑦̅)2
 4. Write an algorithm for Newton’s Iterative Formula to find the 
Reciprocal of a Number N and state the timing Notation 
[𝐻𝑖𝑛𝑡:   𝑥𝑖+1 = 𝑥𝑖(2 – 𝑥𝑖𝑁)] 
Assume x0
 = 0.1 and N = 5; show how global minimum converges at each iteration.